+++ src/SDKs/Compute/Compute.Tests/ScenarioTests/ExtensionTests.cs
@@ -57,7 +57,7 @@ namespace Compute.Tests
 
                     // Delete an extension that does not exist in the VM. A http status code of NoContent should be returned which translates to operation success.
                     m_CrpClient.VirtualMachineExtensions.Delete(rgName, vm.Name, "VMExtensionDoesNotExist");
-                    
+
                     // Add an extension to the VM
                     var vmExtension = GetTestVMExtension();
                     var response = m_CrpClient.VirtualMachineExtensions.CreateOrUpdate(rgName, vm.Name, vmExtension.Name, vmExtension);